+++ src/ft_tcp_socket.h
@@ -9,7 +9,7 @@ struct ft_tcpSocket {
     int port;
 };
 
-struct ft_tcpSocket *ft_tcpSocketCreate(char *host, int port);
+struct ft_tcpSocket *ft_tcpSocketCreate(char *host, unsigned short port);
 int ft_tcpSocketConnect(struct ft_tcpSocket *socket);
 ssize_t ft_tcpSocketReceive(struct ft_tcpSocket *socket, void *buffer, size_t nBytes);
 ssize_t ft_tcpSocketSend(struct ft_tcpSocket *socket, const void *buffer, size_t nBytes);
